etcd-tester: cleanup
